+++ 5/command.hpp
@@ -6,14 +6,14 @@
 class command {
 public:
 
-    command(uint_fast8_t player = 0, const uint_fast8_t place = 0);
-    uint_fast8_t getPlayer();
-    uint_fast8_t getPlace();
+    command(int player = 0, const int place = 0);
+    int getPlayer();
+    int getPlace();
 
 private:
 
-    uint_fast8_t player;
-    const uint_fast8_t place;
+    int player;
+    const int place;
 
 };
